requirements.txt
certifi >= 14.05.14
six >= 1.8.0
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3 >= 1.15.1

Flask >= 1.1.2
flask-restful >= 0.3.9
webargs >= 6.1.0
newsapi-python >= 0.2.6
flasgger >= 0.9.5
